Flush manifest to content layer at the end of markContentAndCreateManifest
Not strictly required for safety, but it makes it easier to reason through
the properties of the error code paths.

For performance reasons, it may (or may not) be desirable to NOT flush
the manifests right away, and instead flush all the manifests together
at the end of a run of the GC-mark phase. This would put all the
manifests in a single content object and may result in better compression
as well since the manifests have the same structure. OTOH, it could be
that in the common case, a single manifest is produced, so there's no
point in trying to batch them.

Anyway, this is simpler.